Move cleanup functions into single helper
