+++ test/integration/query.test.ts
@@ -6,9 +6,8 @@ import Database from '../../src/models/db/index.js'
 import { QueryRow } from '../../src/models/db/types.js'
 import VeritableCloudagent from '../../src/models/veritableCloudagent/index.js'
 import createHttpServer from '../../src/server.js'
-import { cleanupCloudagent } from '../helpers/cloudagent.js'
+import { cleanupCloudagent, cleanupDatabase } from '../helpers/cleanup.js'
 import { TwoPartyConnection, withVerifiedConnection } from '../helpers/connection.js'
-import { cleanupDatabase } from '../helpers/db.js'
 import { bobDbConfig, mockEnvBob } from '../helpers/fixtures.js'
 import { mockLogger } from '../helpers/logger.js'
 import { post } from '../helpers/routeHelper.js'
